Default windows and optimization
Fixed crash when second monitor is disconnected. Now if the SLM monitor is not connected the software falls back to rendering on the main monitor. Optimized code for generation of gratings, now using meshgrid as for lenses. Fixed pattern for lenses instead of cos(X**2+Y**2) now is X**2+Y**2 (as hamamatsu_